@@ -1,9 +1,6 @@
 {
-  "name": "stimulus-content-loader",
+  "name": "@stimulus-components/content-loader",
   "version": "4.2.0",
-  "engines": {
-    "node": "*"
-  },
   "description": "A Stimulus controller to asynchronously load HTML from an url.",
   "keywords": [
     "stimulus",
@@ -21,32 +18,47 @@
   "license": "MIT",
   "homepage": "https://github.com/stimulus-components/stimulus-content-loader",
   "private": false,
+  "publishConfig": {
+    "access": "public"
+  },
   "main": "dist/stimulus-content-loader.umd.js",
   "module": "dist/stimulus-content-loader.mjs",
+  "types": "dist/types/index.d.ts",
   "scripts": {
-    "format": "prettier-standard  '**/*.{ts,css,html}' --format",
-    "lint": "prettier-standard  '**/*.{ts,css,html}' --lint",
+    "lint": "tsc --noEmit && eslint --ext .js,.mjs,.ts . && prettier . --check",
+    "lintfix": "eslint --ext .js,.mjs,.ts . --fix && prettier . --write",
+    "types": "tsc --noEmit false --declaration true --emitDeclarationOnly true --outDir dist/types",
     "dev": "vite",
     "prod": "vite build --mode netlify",
-    "build": "tsc --noEmit && vite build",
+    "build": "vite build && yarn types",
     "version": "yarn build",
-    "np": "np --no-2fa --no-test"
+    "np": "np --no-2fa --no-tests"
   },
   "devDependencies": {
-    "@babel/core": "7.22.10",
-    "@babel/plugin-syntax-class-properties": "7.12.13",
-    "@babel/preset-env": "^7.22.10",
-    "@babel/preset-typescript": "^7.22.5",
-    "@hotwired/stimulus": "^3.2.2",
-    "autoprefixer": "^10.4.14",
-    "np": "^8.0.4",
-    "postcss": "^8.4.27",
-    "prettier-standard": "16.4.1",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^5.1.6",
-    "vite": "^4.4.9"
+    "@babel/core": "7.24.1",
+    "@babel/polyfill": "7.12.1",
+    "@babel/preset-env": "7.24.1",
+    "@babel/preset-typescript": "7.24.1",
+    "@hotwired/stimulus": "^3",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
+    "autoprefixer": "10.4.18",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-standard": "17.1.0",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-n": "16.6.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-standard": "5.0.0",
+    "np": "9.2.0",
+    "postcss": "8.4.37",
+    "prettier": "3.2.5",
+    "tailwindcss": "3.4.1",
+    "typescript": "^5.4.2",
+    "vite": "5.2.0"
   },
   "peerDependencies": {
-    "@hotwired/stimulus": "^3.2.1"
+    "@hotwired/stimulus": "^3.0.0"
   }
 }
