@@ -1,12 +1,15 @@
 // https://eslint.org/docs/user-guide/configuring
 
 module.exports = {
+  root: true,
+  parser: "@typescript-eslint/parser",
   env: {
     browser: true,
     node: true,
-    es6: true
+    es6: true,
+    jest: true,
   },
-  parserOptions: {
-    sourceType: "module"
-  }
+  // https://github.com/standard/standard/blob/master/docs/RULES-en.md
+  extends: ["standard", "prettier"],
+  plugins: ["@typescript-eslint"],
 }
