@@ -1,5 +1,5 @@
-import { resolve } from 'path'
-import { defineConfig } from 'vite'
+import { resolve } from "path"
+import { defineConfig } from "vite"
 
 export default defineConfig(({ mode }) => {
   if (mode === 'netlify') {
@@ -17,20 +17,23 @@ export default defineConfig(({ mode }) => {
   }
 
   return {
+    esbuild: {
+      minifyIdentifiers: false,
+    },
     build: {
       lib: {
-        entry: resolve(__dirname, 'src/index.ts'),
-        name: 'StimulusContentLoader',
-        fileName: 'stimulus-content-loader'
+        entry: resolve(__dirname, "src/index.ts"),
+        name: "StimulusContentLoader",
+        fileName: "stimulus-content-loader",
       },
       rollupOptions: {
-        external: ['@hotwired/stimulus'],
+        external: ["@hotwired/stimulus"],
         output: {
           globals: {
-            '@hotwired/stimulus': 'Stimulus'
-          }
-        }
-      }
-    }
+            "@hotwired/stimulus": "Stimulus",
+          },
+        },
+      },
+    },
   }
 })
